Refactor tile editing logic into partial form
