+++ SPHMMaker/SPHMMaker.csproj
@@ -14,4 +14,8 @@
     <PackageReference Include="System.Data.SqlClient" Version="4.8.6" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Compile Include="..\TileForm.cs" Link="TileForm.cs" />
+  </ItemGroup>
+
 </Project>
